Crate pixels

Modules

decoding 🔒

snapshot 🔒

Structs

ImageFrame

ImageFrameView

A non-owning reference to the data of an ImageFrame

ImageMetadata

RasterImage

SharedRasterImage

A version of RasterImage that can be sent across IPC channels.

SharedSnapshot

A version of Snapshot that can be sent across IPC channels.

Snapshot

Represents image bitmap with metadata, usually as snapshot of canvas

Enums

Alpha

CorsStatus

Whether this response passed any CORS checks, and is thus safe to read from in cross-origin environments.

EncodedImageType

FilterQuality

Multiply

PixelFormat

Repeat

SnapshotAlphaMode

SnapshotData

The data in a Snapshot. If created via shared memory, this will be the SharedMemory variant, but otherwise it is the Owned variant. any attempt to mutate the Snapshot will convert it to the Owned variant.

SnapshotPixelFormat

Functions

clip

compute_rgba8_byte_length_if_within_limit

Computes image byte length, returning None if overflow occurred or the total length exceeds the maximum image allocation size.

copy_rgba8_image

Copies the rectangle of the source image to the destination image.

detect_image_format

flip_y_rgba8_image_inplace

Flips the source image vertically in place.

generic_transform_inplace

is_bmp 🔒

is_gif 🔒

is_ico 🔒

is_jpeg 🔒

is_png 🔒

is_webp 🔒

load_from_memory

multiply_u8_color

Returns a*b/255, rounding any fractional bits to nearest integer to reduce the loss of precision after multiple consequence alpha (un)premultiply operations.

rgba8_byte_swap_and_premultiply_inplace

rgba8_byte_swap_colors_inplace

rgba8_get_rect

rgba8_premultiply_inplace

Returns true if the pixels were found to be completely opaque.

scale_rgba8_image

Scales the source image to the required size, performing sampling filter algorithm.

sensible_delay 🔒

transform_inplace

unmultiply_inplace