pixels

Module snapshot

Structs

SharedSnapshot

A version of Snapshot that can be sent across IPC channels.

Snapshot

Represents image bitmap with metadata, usually as snapshot of canvas

Enums

Alpha

SnapshotAlphaMode

SnapshotData

The data in a Snapshot. If created via shared memory, this will be the SharedMemory variant, but otherwise it is the Owned variant. any attempt to mutate the Snapshot will convert it to the Owned variant.

SnapshotPixelFormat