pixels

Struct RasterImage

pub struct RasterImage {
    pub metadata: ImageMetadata,
    pub format: PixelFormat,
    pub id: Option<ImageKey>,
    pub cors_status: CorsStatus,
    pub bytes: Arc<Vec<u8>>,
    pub frames: Vec<ImageFrame>,
    pub is_opaque: bool,
    pub loop_count: Option<Repeat>,
}

Fields
metadata: ImageMetadataformat: PixelFormatid: Option<ImageKey>cors_status: CorsStatusbytes: Arc<Vec<u8>>frames: Vec<ImageFrame>is_opaque: bool

Whether or not all of the frames of this image are opaque.

loop_count: Option<Repeat>

The loop count for this image’s animation. For animated images, this has a default value of Repeat::Infinite (if no loop count is specified in the image). For images that do not animate, this will be None.

Implementations

impl RasterImage

pub fn should_animate(&self) -> bool

pub(crate) fn frame_view<'image>( &'image self, frame: &ImageFrame, ) -> ImageFrameView<'image>

pub fn frame(&self, index: usize) -> Option<ImageFrameView<'_>>

pub fn first_frame(&self) -> ImageFrameView<'_>

pub fn as_snapshot(&self) -> Snapshot

pub fn frame_data(&self, index: usize) -> Option<&ImageFrame>

pub fn webrender_image_descriptor_and_data_for_frame( &self, frame_index: usize, ) -> (ImageDescriptor, GenericSharedMemory, bool)

Returns tuple containing three items:

An ImageDescriptor descriptor used to describe this image to WebRender

A GenericSharedMemory containing the image data

Whether or not this image should be cached in the Painter animating image cache.

pub fn webrender_image_descriptor_and_offset_for_frame( &self, ) -> Option<ImageDescriptor>

For animations the image already exists in a cache in ‘Painter’. We just send the description. Currently we do not support ‘PixelFormat::RGB8’

pub fn to_shared(&self) -> Arc<SharedRasterImage>