Crate pixels

Structs

Image

ImageMetadata

Enums

CorsStatus

Whether this response passed any CORS checks, and is thus safe to read from in cross-origin environments.

PixelFormat

Functions

clip

detect_image_format

is_bmp 🔒

is_gif 🔒

is_ico 🔒

is_jpeg 🔒

is_png 🔒

is_webp 🔒

load_from_memory

multiply_u8_color

rgba8_byte_swap_and_premultiply_inplace

rgba8_byte_swap_colors_inplace

rgba8_get_rect

rgba8_premultiply_inplace

Returns true if the pixels were found to be completely opaque.

unmultiply_inplace